requirements.txt
redis
fastapi
uvicorn
pyjwt